+++ CHANGELOG.md
@@ -2,9 +2,26 @@
 
 This file is used to list changes made in each version of the win_choco cookbook.
 
-## 0.1.0
+## 0.1.6
+
+- Added NodeJS
+- Added Adobe Acrobat Reader
+
+## 0.1.5
+
+- Added VSCode if Node Attribute
+
+## 0.1.3
+
+- Installs and Validates *Microsoft Edge*
+
+## 0.1.2
+
+- Installs and Validates *7zip*
+
+## 0.1.1
 
 Initial release.
 
-- change 0
-- change 1
+- Installs and Validates *Chocolatey*
+- Installs and Validates *Git*